refactor : method 정보 class 추가
