+++ src/main/java/core/mvc/tobe/HandlerExecution.java
@@ -1,16 +1,11 @@
 package core.mvc.tobe;
 
-import core.annotation.web.RequestMapping;
 import core.mvc.ModelAndView;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.core.LocalVariableTableParameterNameDiscoverer;
-import org.springframework.core.ParameterNameDiscoverer;
-import org.springframework.web.util.pattern.PathPattern;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import java.lang.annotation.Annotation;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 
@@ -27,18 +22,10 @@ public class HandlerExecution {
     }
 
     public ModelAndView handle(HttpServletRequest request, HttpServletResponse response) throws IllegalAccessException, InstantiationException, InvocationTargetException, ClassNotFoundException, NoSuchMethodException {
-        ParameterNameDiscoverer nameDiscoverer = new LocalVariableTableParameterNameDiscoverer();
-        Class<?>[] parameterTypes = method.getParameterTypes();
-        Object[] bindObjects = new Object[parameterTypes.length];
-        String[] names = nameDiscoverer.getParameterNames(method);
-        PathPattern pathPattern = PathPatternUtil.getPathPattern(method.getDeclaredAnnotation(RequestMapping.class).value());
-        Annotation[][] parameterAnnotations = method.getParameterAnnotations();
+        MethodInfo methodInfo = new MethodInfo(method);
+        Object[] bindParameters = methodInfo.bindingParameters(request);
 
-        for (int i = 0; i < parameterTypes.length; i++) {
-            ParameterInfo parameterInfo = new ParameterInfo(parameterTypes[i], names[i], pathPattern, parameterAnnotations[i]);
-            bindObjects[i] = Helpers.executeHelper(parameterInfo, request);
-        }
-        return (ModelAndView) method.invoke(clazz.newInstance(), bindObjects);
+        return (ModelAndView) method.invoke(clazz.newInstance(), bindParameters);
     }
 
 }
